feat(TCOMP_681): split hierarchy on DebouncedParameter and ValueChangedParameter
